Convert double to BigDecomal

+++ app/src/main/java/com/bs/threadsimulator/data/StockInfo.kt
@@ -1,10 +1,12 @@
 package com.bs.threadsimulator.data
 
+import java.math.BigDecimal
+
 data class StockInfo(
     var symbol: String = "",
-    var openingPrice: Double = 0.0,
-    var closingPrice: Double = 0.0,
-    var low: Double = 0.0,
-    var high: Double = 0.0,
-    var currentPrice: Double = 0.0
+    var openingPrice: BigDecimal = BigDecimal(0.0),
+    var closingPrice: BigDecimal = BigDecimal(0.0),
+    var low: BigDecimal = BigDecimal(0.0),
+    var high: BigDecimal = BigDecimal(0.0),
+    var currentPrice: BigDecimal = BigDecimal(0.0)
 )
